moved render task to its own class
