+++ src/index.ts
@@ -1,15 +1,7 @@
 import "../src/sass/main.scss";
 
-import { ChartConfigs } from "./interfaces/ChartConfigs";
-import { SubTask } from "./interfaces/task/SubTask";
 import { Task } from "./interfaces/task/Task";
-import { Calender } from "./libs/HtmlElement/Calender";
-import { createElement } from "./libs/HtmlElement/HtmlHelper";
-import { inputValue } from "./libs/HtmlElement/InputHelper";
-import { Modal } from "./libs/HtmlElement/Modal";
-import { SelectOptionsHelper } from "./libs/HtmlElement/SelectOptionsHelper";
-import { Sidebar } from "./libs/HtmlElement/Sidebar";
-import { Validations } from "./libs/validations";
+import { RenderCalender } from "./libs/HtmlElement/RenderCalender";
 
 // eslint-disable-next-line no-use-before-define
 function isTaskDay(date: Date, tasks: Task[]): boolean {
@@ -19,142 +11,6 @@ function isTaskDay(date: Date, tasks: Task[]): boolean {
 
 		return date >= taskStart && date <= taskEnd;
 	});
-}
-function generateRandomId(length = 16): string {
-	const alphanumeric = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-	let id = "";
-
-	for (let i = 0; i < length; i++) {
-		const randomIndex = Math.floor(Math.random() * alphanumeric.length);
-		id += alphanumeric.charAt(randomIndex);
-	}
-
-	return id;
-}
-
-// Render calendar
-function renderCalendar(configs: ChartConfigs): void {
-	const containerId: string = configs.id;
-	configs.tasks.map((task) => {
-		task.uid = task.uid ?? generateRandomId();
-
-		return task;
-	});
-	const tasks: Task[] = configs.tasks;
-	const container = document.getElementById(containerId);
-	if (!container) {
-		return;
-	}
-	container.classList.add("container");
-
-	const sideBarEl = createElement("div", "sidebar");
-	sideBarEl.id = "sidebar";
-	container.appendChild(sideBarEl);
-
-	const mainBox = createElement("div", "mainBox");
-	container.appendChild(mainBox);
-	mainBox.id = "mainBox";
-	// Render row with day numbers
-	const calendar: Calender = new Calender(mainBox, tasks);
-	const sidebar: Sidebar = new Sidebar(sideBarEl, tasks);
-	calendar.renderCalender();
-	sidebar.renderTasksSidebar();
-	const inputs = configs.modalConfigs.inputs;
-	// setTasksInputs(inputs, tasks);
-	const taskOptions = tasks
-		.filter((task) => task.uid)
-		.map((task) => ({ label: task.name, value: task.uid }));
-	new SelectOptionsHelper(inputs).setTasksInputs("parentTask", taskOptions);
-	new SelectOptionsHelper(inputs).setTasksInputs("successor", taskOptions);
-	new SelectOptionsHelper(inputs).setTasksInputs("predecessor", taskOptions);
-	Modal.renderModal(configs.modalConfigs);
-	const addTask = document.getElementById("addTask");
-	if (addTask) {
-		addTask.addEventListener("click", () => {
-			const mainBox = document.getElementById("mainBox");
-			const task: Task = {
-				name: inputValue("name") ?? "",
-				start: inputValue("start") ?? "",
-				end: inputValue("end") ?? "",
-				parentTask: inputValue("parentTask") ?? "",
-				completion: inputValue("completion") ?? "",
-				uid: generateRandomId(),
-			};
-			document.querySelectorAll("span.error").forEach((element) => {
-				element.remove();
-			});
-			const validation = new Validations(task, inputs);
-			validation.errors().forEach((error) => {
-				const errorMessage = error.messages;
-				const errorEl = createElement("span", "error", errorMessage.join());
-				const inputEl = document.querySelector(`[name='${error.field}']`) as HTMLElement;
-				if (inputEl.parentNode) {
-					inputEl.parentNode.appendChild(errorEl);
-				}
-			});
-			if (validation.errors().length === 0) {
-				const elementsToRemove = document.querySelectorAll(".task-row");
-				elementsToRemove.forEach((element) => {
-					element.remove();
-				});
-				const uid = document.querySelector("input[name='_uid']") as HTMLInputElement | undefined;
-				if (!task.parentTask) {
-					if (uid) {
-						configs.tasks = configs.tasks.map((taskObj) => {
-							if (taskObj.uid === uid.value) {
-								return { ...taskObj, ...task };
-							}
-
-							return taskObj;
-						});
-					} else {
-						configs.tasks.push(task);
-					}
-				} else {
-					const parentTaskIndex = configs.tasks.findIndex(
-						(taskObj) => taskObj.uid === task.parentTask
-					);
-					const parentTask = configs.tasks[parentTaskIndex];
-					if (uid) {
-						configs.tasks[parentTaskIndex].subTasks = parentTask.subTasks?.map((taskObj: SubTask) => {
-								if (taskObj.uid === uid.value) {
-									return { ...taskObj, ...task };
-								}
-
-								return taskObj;
-							}
-						);
-					} else if (parentTask.subTasks) {
-						parentTask.subTasks.push(<SubTask>task);
-					} else {
-						parentTask["subTasks"] = [];
-						parentTask.subTasks.push(<SubTask>task);
-					}
-				}
-
-				calendar.updateTasks(configs.tasks);
-				sidebar.updateTasks(configs.tasks);
-
-				Modal.closeModal();
-				const taskOptions = tasks
-					.filter((task) => task.uid)
-					.map((task) => ({ label: task.name, value: task.uid }));
-				new SelectOptionsHelper(inputs).setTasksInputs("parentTask", taskOptions);
-				new SelectOptionsHelper(inputs).setTasksInputs("successor", taskOptions);
-				new SelectOptionsHelper(inputs).setTasksInputs("predecessor", taskOptions);
-				const form = document.querySelector("form#taskForm") as HTMLFormElement;
-				form.reset();
-				if (mainBox) {
-					if (configs.modalConfigs.addTask) {
-						configs.modalConfigs.addTask(task);
-					}
-					if (configs.modalConfigs.updateTask) {
-						configs.modalConfigs.updateTask(task);
-					}
-				}
-			}
-		});
-	}
 }
 
 // Get the current date
@@ -211,7 +67,7 @@ const tasks = [
 		completion: 0,
 	},
 ];
-renderCalendar({
+new RenderCalender({
 	id: "renderCalendar",
 	tasks,
 	modalConfigs: {
@@ -270,7 +126,7 @@ renderCalendar({
 			},
 		],
 	},
-});
+}).renderCalendar();
 function addTask(task: Task) {
 	console.log(task, "for user");
 }
